refactor: Moved DateSelection from place-screen to date-picker, and renamed to date-selection
