--- src/date-picker/index.ts
@@ -1,2 +1,0 @@
-export {DatePickerSheet} from './DatePickerSheet';
-export type {DateOptionAndValue} from './types';
