--- src/date-picker/types.ts
@@ -1,9 +1,0 @@
-export type DateOptionAndText<T extends string> = {
-  option: T;
-  text: string;
-  selected?: boolean;
-};
-export type DateOptionAndValue<T extends string> = {
-  option: T;
-  date: string;
-};
