+++ src/screen-components/place-screen/components/QuayView.tsx
@@ -1,10 +1,8 @@
 import {Quay, StopPlace} from '@atb/api/types/departures';
-import {useFavoritesContext} from '@atb/modules/favorites';
-import {DepartureSearchTime} from '../types';
 import {StyleSheet} from '@atb/theme';
 import React, {useEffect} from 'react';
 import {RefreshControl, SectionList, SectionListData, View} from 'react-native';
-import {DateSelection} from './DateSelection';
+import {DateSelection} from '@atb/components/date-selection';
 import {FavoriteToggle} from './FavoriteToggle';
 import {QuaySection} from './QuaySection';
 import {useDeparturesData} from '../hooks/use-departures-data';
@@ -14,6 +12,8 @@ import {MessageInfoBox} from '@atb/components/message-info-box';
 import {DeparturesTexts, dictionary, useTranslation} from '@atb/translations';
 import {useIsFocused} from '@react-navigation/native';
 import type {ContrastColor} from '@atb-as/theme';
+import type {DepartureSearchTime} from 'src/components/date-selection';
+import {useFavoritesContext} from '@atb/modules/favorites';
 
 const NUMBER_OF_DEPARTURES_PER_QUAY_TO_SHOW = 1000;
