+++ src/screen-components/place-screen/index.tsx
@@ -1,6 +1,5 @@
 export {PlaceScreenComponent} from './PlaceScreenComponent';
 export {StopPlacesView} from './components/StopPlacesView';
-export type {DepartureSearchTime} from './types';
 export type {PlaceScreenParams} from './PlaceScreenComponent';
-
 export {useStopsDetailsDataQuery} from './hooks/use-stops-details-data-query';
+export type {StopPlaceAndQuay} from './types';
